\name{analyze.sequences}
\alias{analyze.sequences}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
analyzes the fit of DNA sequences to gene trees estimated by *BEAST. 
}
\description{
takes data read into R using read.starbeast and uses posterior predictive simulation to evaluate the fit of the sequence data to the estimated gene trees from the analysis
}
\usage{
analyze.sequences(data, seqgendir = "/directory/containing/seqgen")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
data read in using read.starbeast
}
  \item{seqgendir}{
seqgendir is the directory containing the a compiled executable of Rambaut's seq-gen software. this software is used to simulate DNA sequence data. 
}
}
\details{
This function analyzes the fit of DNA sequence data to the estimated gene trees generated by *BEAST. It does this by simulating new sequence data from the posterior and comparing the simulated and empirical sequence data. It uses four test statistics: the number of variable sites, the phylogenetic likelihood of the data given the gene tree, the multinomial likelihood of the sequence data and the GC statistic (the difference between the multinomial and phylogenetic likelihoods). It calculates these statistics for each locus (or each subset of nucleotides if a locus is partitioned). The results can be summarized using 'summary' or visualized using 'plot'. 
}
\value{
The function returns an object of class "sequenceteststats". This object contains three lists, "empirical" and "simulated" and "test.stats", which contain the empirical ("empirical") and simulated ("simulated") values and the difference between them ("test.stat"), respectively. "empirical" contains 3 matrices and one vector containing the test statistics for the empirical data. "simulated" contains 4 matrices with those test statistics. "test.stat" contains 4 matrices with the differences between simulated and empirical. Those differences give a test distribution whose expectation is 0. The four elements of each list are called "phy.likelihood", "multi.likelihood", "gcstat" and "v.sites"
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Noah Reid
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
