\name{read.gtree.samples}
\alias{read.gtree.samples}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
reads gene tree samples from the posterior generated by *BEAST
}
\description{
this function will read in a .trees file generated by *BEAST and store the among-branch rate variation in a vector called "rate" appended to each tree}
\usage{
read.gtree.samples(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
the path and file name of a .trees nexus file generated by *BEAST corresponding to the posterior distribution of a coalescent genealogy for a single locus. 
}
}
\details{
this function uses ape's read.nexus and another function, taken and reconstituted from package 'phyloch' by Christoph Heibl to get the trees and branch rate scalar vector from the posterior distribution of gene trees from a *BEAST run. 

this function is unfortunately very slow compared to read.nexus. 
}
\value{
a phylo or multiphylo object, each with a vector of among-branch rate variation named "rate" appended. 
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Noah Reid}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
